fix: resolve NPE when attempting to use a null modelClass by trying to use the x-parser-schema-id property of the source schemafirst, then try the property name.

+++ hooks/post-process.js
@@ -2,6 +2,7 @@
 const fs = require('fs');
 const path = require('path');
 const ApplicationModel = require('../lib/applicationModel.js');
+const _ = require('lodash');
 const applicationModel = new ApplicationModel('post');
 // To enable debug logging, set the env var DEBUG="postProcess" with whatever things you want to see.
 const debugPostProcess = require('debug')('postProcess');
@@ -74,12 +75,12 @@ function processSchema(generator, schemaName, schema, sourcePath, defaultJavaPac
   const filePath = path.resolve(sourcePath, fileName);
   debugPostProcess(`processSchema ${schemaName}`);
   debugPostProcess(schema);
-  if (schema.type() !== 'object') {
+  const modelClass = applicationModel.getModelClass({schema: schema, schemaName: schemaName});
+  const javaName = modelClass.getClassName();
+  if (schema.type() !== 'object' || _.startsWith(javaName, 'Anonymous')) {
     debugPostProcess(`deleting ${filePath}`);
-    fs.unlinkSync(filePath);
+    fs.existsSync(filePath) && fs.unlinkSync(filePath);
   } else {
-    const modelClass = applicationModel.getModelClass(schemaName);
-    const javaName = modelClass.getClassName();
     const packageDir = getPackageDir(generator, defaultJavaPackageDir, modelClass);
     debugPostProcess(`packageDir: ${packageDir}`);
